requirements.txt
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
gunicorn==20.1.0
Flask==2.1.2
Werkzeug==2.0.3